gmball: code refactor
